Successful game redesign
I changed the game from using a full drum kit, to a selection of parts. The player now navigates the drum kit using the yaw axis rotation of the controllers. To play a part of the drum kit, once the drumstick is in a certain area of the drum kit, the player must squeeze the controller to play that part.

Removed any unnecessary scripts.